split the pkg/component package
This package is used in the production code, but most of it is used in
build time, to create the manifests.

This creates coupling that make it hard to maintain this pretty big
package.

This PR takes the parts that are only relevant for the manifest
generation, out to the new `tools/manifests` package, leaving only a
small part of the original package, that used by the production or the
test code.
